refactor: split feishu sync tests
