+++ tests/feishu_sync.test.js
@@ -1,7 +1,7 @@
 /*
 ## 核心功能
 
-覆盖 feishu sync 相关行为的 Vitest 测试用例。
+覆盖 feishu sync 基础服务 相关行为的 Vitest 测试用例。
 
 ## 输入
 
@@ -9,7 +9,7 @@
 
 ## 输出
 
-输出自动化断言结果，保护渲染、同步、设置、安全或 UI 行为不回归。
+输出自动化断言结果，保护Markdown 处理、设置归一化、块转换和 API 客户端行为不回归。
 
 ## 定位
 
@@ -17,7 +17,7 @@
 
 ## 依赖
 
-关键依赖：Vitest、项目 mock/helper，以及被测的 feishu sync 模块。
+关键依赖：Vitest、项目 mock/helper，以及被测的飞书同步模块。
 
 ## 维护规则
 
@@ -28,8 +28,9 @@
 import { describe, it, expect, vi, beforeEach, beforeAll } from 'vitest';
 const obsidianMock = require('obsidian');
 
-// Mock requestUrl BEFORE requiring modules under test so they destructure the spy
-obsidianMock.requestUrl = vi.fn();
+if (!obsidianMock.requestUrl || typeof obsidianMock.requestUrl.mockReset !== 'function') {
+  obsidianMock.requestUrl = vi.fn();
+}
 
 let FeishuApiClient;
 let stripYamlFrontmatter;
@@ -37,9 +38,6 @@ let parseYamlTitle;
 let convertWikilinks;
 let convertObsidianImageSyntax;
 let extractImagesFromMarkdown;
-let prepareLocalImagesForFeishu;
-let prepareMermaidDiagramsForFeishu;
-let syncNoteToFeishu;
 let createDefaultFeishuSyncSettings;
 let FEISHU_FREE_MONTHLY_API_LIMIT;
 let normalizeFeishuSyncSettings;
@@ -53,24 +51,6 @@ let removeFeishuMermaidPreferenceByPath;
 let getFeishuDirectChildBlocks;
 let summarizeFeishuBlockChunk;
 let buildFeishuCreatePayloadBlocks;
-
-function makePngBytes(width = 100, height = 50) {
-  const bytes = new Uint8Array(24);
-  bytes.set([0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a], 0);
-  const view = new DataView(bytes.buffer);
-  view.setUint32(16, width, false);
-  view.setUint32(20, height, false);
-  return bytes.buffer;
-}
-
-function makeGifBytes(width = 100, height = 50) {
-  const bytes = new Uint8Array(10);
-  bytes.set([0x47, 0x49, 0x46, 0x38, 0x39, 0x61], 0);
-  const view = new DataView(bytes.buffer);
-  view.setUint16(6, width, true);
-  view.setUint16(8, height, true);
-  return bytes.buffer;
-}
 
 beforeAll(async () => {
   globalThis.obsidian = obsidianMock;
@@ -86,11 +66,9 @@ beforeAll(async () => {
   extractImagesFromMarkdown = processorMod.extractImagesFromMarkdown;
 
   const syncMod = await import('../services/feishu-sync.js');
-  prepareLocalImagesForFeishu = syncMod.prepareLocalImagesForFeishu;
-  syncNoteToFeishu = syncMod.syncNoteToFeishu;
-
-  const mermaidMod = await import('../services/feishu-mermaid-renderer.js');
-  prepareMermaidDiagramsForFeishu = mermaidMod.prepareMermaidDiagramsForFeishu;
+  getFeishuDirectChildBlocks = syncMod.getFeishuDirectChildBlocks;
+  summarizeFeishuBlockChunk = syncMod.summarizeFeishuBlockChunk;
+  buildFeishuCreatePayloadBlocks = syncMod.buildFeishuCreatePayloadBlocks;
 
   const settingsMod = await import('../services/feishu-settings.js');
   createDefaultFeishuSyncSettings = settingsMod.createDefaultFeishuSyncSettings;
@@ -103,9 +81,6 @@ beforeAll(async () => {
   getFeishuMermaidPreferenceByPath = settingsMod.getFeishuMermaidPreferenceByPath;
   setFeishuMermaidPreferenceByPath = settingsMod.setFeishuMermaidPreferenceByPath;
   removeFeishuMermaidPreferenceByPath = settingsMod.removeFeishuMermaidPreferenceByPath;
-  getFeishuDirectChildBlocks = syncMod.getFeishuDirectChildBlocks;
-  summarizeFeishuBlockChunk = syncMod.summarizeFeishuBlockChunk;
-  buildFeishuCreatePayloadBlocks = syncMod.buildFeishuCreatePayloadBlocks;
 });
 
 describe('Feishu Markdown Processor', () => {
@@ -624,1756 +599,5 @@ describe('Feishu Api Client', () => {
       member_id: 'ou-user-123',
       member_type: 'userid',
     });
-  });
-});
-
-describe('Feishu Sync Coordinator', () => {
-  let app;
-  let settings;
-  let activeFile;
-
-  beforeEach(() => {
-    settings = createDefaultFeishuSyncSettings();
-    settings.appId = 'app-id';
-    settings.appSecret = 'app-secret';
-    settings.folderToken = 'folder-token';
-
-    activeFile = {
-      path: 'notes/test-note.md',
-      basename: 'test-note',
-    };
-
-    app = {
-      metadataCache: {
-        getFirstLinkpathDest: vi.fn(),
-      },
-      vault: {
-        readBinary: vi.fn(),
-      },
-    };
-
-    obsidianMock.requestUrl.mockReset();
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { json: { code: 0, data: { ticket: 'ticket_123' } } };
-        } else {
-          return { json: { code: 0, data: { result: { job_status: 0, token: 'doc_token_456', url: 'https://feishu.cn/docx/doc_token_456' } } } };
-        }
-      }
-      if (url.includes('permissions') && url.includes('transfer_owner')) {
-        return { json: { code: 0, msg: 'success' } };
-      }
-      if (url.includes('files?folder_token')) {
-        return { json: { code: 0, data: { files: [] } } };
-      }
-      return { json: { code: 0 } };
-    });
-  });
-
-  it('should prepare local Markdown images as Feishu replacement assets', async () => {
-    const localFile = {
-      path: 'notes/attachments/local.png',
-      name: 'local.png',
-      extension: 'png',
-      bytes: makePngBytes(100, 50),
-    };
-    app.metadataCache.getFirstLinkpathDest.mockImplementation((linkpath) => (
-      linkpath === 'attachments/local.png' ? localFile : null
-    ));
-    app.vault.getAbstractFileByPath = vi.fn((filePath) => (
-      filePath === 'notes/attachments/local.png' ? localFile : null
-    ));
-    app.vault.getResourcePath = vi.fn(() => 'app://local/local.png');
-    app.vault.readBinary.mockImplementation(async (file) => file.bytes);
-
-    const result = await prepareLocalImagesForFeishu(
-      app,
-      activeFile,
-      '# Test\n![Local](attachments/local.png)\n![Remote](https://cdn.example.com/a.png)'
-    );
-
-    expect(result.warnings).toEqual([]);
-    expect(result.assets).toHaveLength(1);
-    expect(result.assets[0]).toMatchObject({
-      filename: 'local.png',
-      mimeType: 'image/png',
-      source: {
-        vaultRelativePath: 'notes/attachments/local.png',
-      },
-    });
-    expect(result.assets[0].base64).toBeUndefined();
-    expect(result.markdown).toContain('![Local](https://obsidian-wechat-converter.invalid/feishu-local-image/image-1.png)');
-    expect(result.markdown).toContain('![Remote](https://cdn.example.com/a.png)');
-  });
-
-  it('should prepare Obsidian wiki image embeds with Chinese paths for Feishu replacement', async () => {
-    const localFile = {
-      path: 'notes/attachments/音乐卡点调整.png',
-      name: '音乐卡点调整.png',
-      extension: 'png',
-      bytes: makePngBytes(100, 50),
-    };
-    app.metadataCache.getFirstLinkpathDest.mockReturnValue(localFile);
-    app.vault.getAbstractFileByPath = vi.fn(() => null);
-    app.vault.getResourcePath = vi.fn(() => 'app://local/music.png');
-    app.vault.readBinary.mockImplementation(async (file) => file.bytes);
-
-    const result = await prepareLocalImagesForFeishu(
-      app,
-      activeFile,
-      '![[attachments/音乐卡点调整.png|音乐|510]]'
-    );
-
-    expect(result.warnings).toEqual([]);
-    expect(result.assets[0]).toMatchObject({
-      filename: '音乐卡点调整.png',
-      source: {
-        vaultRelativePath: 'notes/attachments/音乐卡点调整.png',
-      },
-    });
-    expect(result.assets[0].base64).toBeUndefined();
-    expect(result.markdown).toBe('![音乐](https://obsidian-wechat-converter.invalid/feishu-local-image/image-1.png)');
-  });
-
-  it('should prepare local GIF files as Feishu replacement assets', async () => {
-    const gifFile = {
-      path: 'notes/attachments/demo.gif',
-      name: 'demo.gif',
-      extension: 'gif',
-      bytes: makeGifBytes(320, 180),
-    };
-    app.metadataCache.getFirstLinkpathDest.mockReturnValue(gifFile);
-    app.vault.getAbstractFileByPath = vi.fn(() => gifFile);
-    app.vault.getResourcePath = vi.fn(() => 'app://local/demo.gif');
-    app.vault.readBinary.mockImplementation(async (file) => file.bytes);
-
-    const result = await prepareLocalImagesForFeishu(
-      app,
-      activeFile,
-      '![Gif](attachments/demo.gif)'
-    );
-
-    expect(result.warnings).toEqual([]);
-    expect(result.assets).toHaveLength(1);
-    expect(result.assets[0]).toMatchObject({
-      filename: 'demo.gif',
-      mimeType: 'image/gif',
-      source: {
-        vaultRelativePath: 'notes/attachments/demo.gif',
-      },
-    });
-    expect(result.assets[0].base64).toBeUndefined();
-    expect(result.markdown).toBe('![Gif](https://obsidian-wechat-converter.invalid/feishu-local-image/image-1.gif)');
-    expect(app.vault.readBinary).toHaveBeenCalledTimes(1);
-  });
-
-  it('should prepare Mermaid fences as Feishu image placeholder assets', async () => {
-    const renderMermaidFenceToDataUrl = vi.fn(async () => 'data:image/png;base64,bWVybWFpZA==');
-
-    const result = await prepareMermaidDiagramsForFeishu(
-      'Before\n```mermaid\ngraph TD\nA-->B\n```\nAfter',
-      {
-        renderMermaidFenceToDataUrl,
-        localImageSrcFactory: (asset) => `https://obsidian-wechat-converter.invalid/feishu-local-image/${asset.id}.png`,
-        notePath: activeFile.path,
-      }
-    );
-
-    expect(result.warnings).toEqual([]);
-    expect(result.assets).toHaveLength(1);
-    expect(result.assets[0]).toMatchObject({
-      id: 'feishu-mermaid-1',
-      filename: 'mermaid-diagram-1.png',
-      mimeType: 'image/png',
-      base64: 'bWVybWFpZA==',
-    });
-    expect(result.markdown).toContain('![Mermaid diagram 1](https://obsidian-wechat-converter.invalid/feishu-local-image/feishu-mermaid-1.png)');
-    expect(result.markdown).not.toContain('```mermaid');
-  });
-
-  it('should keep Mermaid source when local rendering is unavailable', async () => {
-    const result = await prepareMermaidDiagramsForFeishu(
-      '```mermaid\ngraph TD\nA-->B\n```',
-      { mermaidApi: null }
-    );
-
-    expect(result.assets).toEqual([]);
-    expect(result.markdown).toContain('```mermaid');
-    expect(result.warnings[0]).toMatchObject({
-      code: 'feishu_mermaid_render_unavailable',
-      severity: 'info',
-    });
-  });
-
-  it('should keep Mermaid source when an explicitly injected renderer fails', async () => {
-    const renderMermaidFenceToDataUrl = vi.fn(async () => {
-      throw new Error('renderer unavailable');
-    });
-
-    const result = await prepareMermaidDiagramsForFeishu(
-      '```mermaid\ngraph TD\nA-->B\n```',
-      { renderMermaidFenceToDataUrl }
-    );
-
-    expect(renderMermaidFenceToDataUrl).toHaveBeenCalledTimes(1);
-    expect(result.assets).toEqual([]);
-    expect(result.markdown).toContain('```mermaid');
-    expect(result.warnings[0]).toMatchObject({
-      code: 'feishu_mermaid_render_failed',
-      severity: 'warning',
-    });
-  });
-
-  it('should leave missing local images unchanged and report a warning', async () => {
-    app.metadataCache.getFirstLinkpathDest.mockReturnValue(null);
-    app.vault.getAbstractFileByPath = vi.fn(() => null);
-
-    const result = await prepareLocalImagesForFeishu(
-      app,
-      activeFile,
-      '![Missing](attachments/missing.png)'
-    );
-
-    expect(result.markdown).toBe('![Missing](attachments/missing.png)');
-    expect(result.warnings.map((warning) => warning.code)).toEqual(['image_local_missing']);
-  });
-
-  it('should import new note successfully', async () => {
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Test Note\nSome text.',
-    });
-
-    expect(result.docToken).toBe('doc_token_456');
-    expect(result.url).toBe('https://feishu.cn/docx/doc_token_456');
-    expect(settings.uploadHistory.length).toBe(1);
-    expect(settings.uploadHistory[0].docToken).toBe('doc_token_456');
-  });
-
-  it('should keep Mermaid source during Feishu sync instead of rasterizing in Obsidian renderer', async () => {
-    const mermaidApi = {
-      render: vi.fn(async () => ({
-        svg: '<svg id="feishu-mermaid-sync" viewBox="0 0 120 80"><rect width="120" height="80"></rect></svg>',
-      })),
-    };
-    const rasterizeSvg = vi.fn(async () => ({
-      dataUrl: 'data:image/png;base64,bWVybWFpZA==',
-      width: 120,
-      height: 80,
-      style: '',
-    }));
-
-    await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Diagram\n```mermaid\ngraph TD\nA-->B\n```',
-      mermaidApi,
-      rasterizeSvg,
-    });
-
-    const uploadCall = obsidianMock.requestUrl.mock.calls.find((call) => (
-      call[0].url.includes('files/upload_all') && call[0].method === 'POST'
-    ));
-    const bodyText = new TextDecoder().decode(new Uint8Array(uploadCall[0].body));
-    expect(mermaidApi.render).not.toHaveBeenCalled();
-    expect(rasterizeSvg).not.toHaveBeenCalled();
-    expect(bodyText).toContain('```mermaid');
-    expect(bodyText).toContain('graph TD');
-  });
-
-  it('should render Mermaid as Feishu image assets only when remote-image mode is selected', async () => {
-    const renderMermaidFenceToDataUrl = vi.fn(async () => 'data:image/png;base64,bWVybWFpZA==');
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { status: 200, json: { code: 0, data: { ticket: 'ticket_123' } } };
-        }
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'doc_token_456', url: 'https://feishu.cn/docx/doc_token_456' } } } };
-      }
-      if (url.includes('blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [{ block_id: 'image_block_1', parent_id: 'doc_token_456', block_type: 27 }],
-            },
-          },
-        };
-      }
-      if (url.includes('medias/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'mermaid_image_token' } } };
-      }
-      if (url.includes('/blocks/image_block_1')) {
-        return { status: 200, json: { code: 0, data: {} } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Diagram\n```mermaid\ngraph TD\nA-->B\n```',
-      mermaidRenderMode: 'remote-image',
-      renderMermaidFenceToDataUrl,
-    });
-
-    expect(renderMermaidFenceToDataUrl).toHaveBeenCalledTimes(1);
-    expect(result.imageSummary).toMatchObject({
-      uploaded: 1,
-      skipped: 0,
-      failed: 0,
-    });
-
-    const calls = obsidianMock.requestUrl.mock.calls.map((call) => call[0]);
-    const markdownUpload = calls.find((request) => request.url.includes('files/upload_all'));
-    const markdownBody = new TextDecoder().decode(new Uint8Array(markdownUpload.body));
-    expect(markdownBody).toContain('![Mermaid diagram 1](https://obsidian-wechat-converter.invalid/feishu-local-image/feishu-mermaid-1.png)');
-    expect(markdownBody).not.toContain('```mermaid');
-
-    const imageUpload = calls.find((request) => request.url.includes('medias/upload_all'));
-    expect(imageUpload).toBeTruthy();
-    const imageUploadBody = new TextDecoder().decode(new Uint8Array(imageUpload.body));
-    expect(imageUploadBody).toContain('filename="mermaid-diagram-1.png"');
-    expect(imageUploadBody).toContain('Content-Type: image/png');
-  });
-
-  it('should keep Mermaid source and continue sync when remote rendering fails', async () => {
-    const renderMermaidFenceToDataUrl = vi.fn(async () => {
-      throw new Error('remote renderer unavailable');
-    });
-    const warnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {});
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Diagram\n```mermaid\ngraph TD\nA-->B\n```',
-      mermaidRenderMode: 'remote-image',
-      renderMermaidFenceToDataUrl,
-    });
-
-    expect(result.docToken).toBe('doc_token_456');
-    expect(result.imageSummary.skipped).toBe(1);
-    expect(result.imageSummary.details[0]).toMatchObject({
-      filename: 'mermaid-diagram-1.png',
-      status: 'skipped',
-      reason: 'feishu_mermaid_render_failed',
-    });
-
-    const uploadCall = obsidianMock.requestUrl.mock.calls.find((call) => (
-      call[0].url.includes('files/upload_all') && call[0].method === 'POST'
-    ));
-    const bodyText = new TextDecoder().decode(new Uint8Array(uploadCall[0].body));
-    expect(bodyText).toContain('```mermaid');
-    expect(bodyText).toContain('graph TD');
-
-    warnSpy.mockRestore();
-  });
-
-  it('should use the Obsidian file basename as the default Feishu document title', async () => {
-    await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# 一级标题\nSome text.',
-    });
-
-    const importTaskRequest = obsidianMock.requestUrl.mock.calls.find((call) => (
-      call[0].url.includes('import_tasks') && call[0].method === 'POST'
-    ))[0];
-    expect(JSON.parse(importTaskRequest.body).file_name).toBe('test-note');
-  });
-
-  it('should perform smart update for previously synced note via temporary imported document order', async () => {
-    settings.uploadHistory = [{
-      title: 'test-note',
-      url: 'https://feishu.cn/docx/doc_token_456',
-      docToken: 'doc_token_456',
-      sourcePath: 'notes/test-note.md',
-      uploadTime: '2026-06-19T00:00:00Z',
-    }];
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('blocks?page_size')) {
-        if (url.includes('/documents/doc_token_456/blocks?page_size')) {
-          return {
-            json: {
-              code: 0,
-              data: {
-                items: [
-                  { block_id: 'doc_token_456', parent_id: '', block_type: 1, children: ['old_child_1'] },
-                  { block_id: 'old_child_1', parent_id: 'doc_token_456', block_type: 2, text: { elements: [{ text_run: { content: 'old' } }] } },
-                ]
-              }
-            }
-          };
-        }
-        return {
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'temp_doc_token', parent_id: '', block_type: 1, children: ['temp_heading', 'temp_paragraph'] },
-                { block_id: 'temp_heading', parent_id: 'temp_doc_token', block_type: 3, heading1: { elements: [{ text_run: { content: 'Updated title' } }] } },
-                { block_id: 'temp_paragraph', parent_id: 'temp_doc_token', block_type: 2, text: { elements: [{ text_run: { content: 'Updated content.' } }] } },
-              ]
-            }
-          }
-        };
-      }
-      if (url.includes('/blocks/doc_token_456/children/batch_delete') && options.method === 'DELETE') {
-        return { json: { code: 0 } };
-      }
-      if (url.includes('import_tasks') && options.method === 'POST') {
-        return { json: { code: 0, data: { ticket: 'ticket_temp_update' } } };
-      }
-      if (url.includes('/import_tasks/ticket_temp_update')) {
-        return { json: { code: 0, data: { result: { job_status: 0, token: 'temp_doc_token', url: 'https://feishu.cn/docx/temp_doc_token' } } } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('children?document_revision_id')) {
-        return { json: { code: 0 } };
-      }
-      if (url.includes('/drive/v1/files/temp_doc_token?type=docx') && options.method === 'DELETE') {
-        return { json: { code: 0 } };
-      }
-      if (url.includes('/drive/v1/files/temp_file_token?type=file') && options.method === 'DELETE') {
-        return { json: { code: 0 } };
-      }
-      return { json: { code: 0 } };
-    });
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# test-note\nUpdated content.',
-    });
-
-    expect(result.docToken).toBe('doc_token_456');
-    expect(settings.uploadHistory.length).toBe(1);
-    const calls = obsidianMock.requestUrl.mock.calls;
-    const deleteCall = calls.find(c => c[0].url.includes('/blocks/doc_token_456/children/batch_delete') && c[0].method === 'DELETE');
-    expect(deleteCall).toBeTruthy();
-    expect(JSON.parse(deleteCall[0].body)).toEqual({ start_index: 0, end_index: 1 });
-    const createCall = calls.find(c => c[0].url.includes('/documents/doc_token_456/blocks/doc_token_456/children?document_revision_id=-1'));
-    expect(JSON.parse(createCall[0].body)).toEqual({
-      index: 1,
-      children: [
-        { block_type: 3, heading1: { elements: [{ text_run: { content: 'Updated title' } }] } },
-        { block_type: 2, text: { elements: [{ text_run: { content: 'Updated content.' } }] } },
-      ],
-    });
-    expect(calls.some(c => c[0].url.includes('children?document_revision_id'))).toBe(true);
-    expect(calls.some(c => c[0].url.includes('transfer_owner'))).toBe(false);
-    expect(calls.some(c => c[0].url.includes('/drive/v1/files/temp_doc_token?type=docx') && c[0].method === 'DELETE')).toBe(true);
-  });
-
-  it('should build create payload from imported document root children order', async () => {
-    settings.uploadHistory = [{
-      title: 'test-note',
-      url: 'https://feishu.cn/docx/doc_token_456',
-      docToken: 'doc_token_456',
-      sourcePath: 'notes/test-note.md',
-      uploadTime: '2026-06-19T00:00:00Z',
-    }];
-
-    let rootCreateCount = 0;
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('blocks?page_size')) {
-        if (url.includes('/documents/doc_token_456/blocks?page_size')) {
-          return {
-            status: 200,
-            json: {
-              code: 0,
-              data: {
-                items: [
-                  { block_id: 'doc_token_456', parent_id: '', block_type: 1, children: ['old_child_1'] },
-                  { block_id: 'old_child_1', parent_id: 'doc_token_456', block_type: 2 },
-                ],
-              },
-            },
-          };
-        }
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'temp_doc_token', parent_id: '', block_type: 1, children: ['bullet_1', 'paragraph_2'] },
-                {
-                  block_id: 'bullet_1',
-                  parent_id: 'temp_doc_token',
-                  block_type: 13,
-                  ordered: { elements: [{ text_run: { content: '第一步' } }], style: { align: 1 } },
-                  children: ['paragraph_1'],
-                },
-                {
-                  block_id: 'paragraph_1',
-                  parent_id: 'bullet_1',
-                  block_type: 12,
-                  bullet: { elements: [{ text_run: { content: 'item 1' } }], style: { align: 1 } },
-                },
-                {
-                  block_id: 'paragraph_2',
-                  parent_id: 'temp_doc_token',
-                  block_type: 2,
-                  text: { elements: [{ text_run: { content: 'tail' } }] },
-                },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('import_tasks') && options.method === 'POST') {
-        return { status: 200, json: { code: 0, data: { ticket: 'ticket_temp_update' } } };
-      }
-      if (url.includes('/import_tasks/ticket_temp_update')) {
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'temp_doc_token', url: 'https://feishu.cn/docx/temp_doc_token' } } } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('/documents/doc_token_456/blocks/created_ordered_1/children?document_revision_id=-1')) {
-        const body = JSON.parse(options.body);
-        expect(body).toEqual({
-          index: 0,
-          children: [
-            {
-              block_type: 12,
-              bullet: { elements: [{ text_run: { content: 'item 1' } }], style: { align: 1 } },
-            },
-          ],
-        });
-        return { status: 200, json: { code: 0, data: { children: [{ block_id: 'created_bullet_1' }] } } };
-      }
-      if (url.includes('/documents/doc_token_456/blocks/doc_token_456/children?document_revision_id=-1') && options.method === 'POST') {
-        const body = JSON.parse(options.body);
-        if (rootCreateCount === 0) {
-          expect(body).toEqual({
-            index: 1,
-            children: [
-              {
-                block_type: 13,
-                ordered: { elements: [{ text_run: { content: '第一步' } }], style: { align: 1 } },
-              },
-            ],
-          });
-          rootCreateCount += 1;
-          return { status: 200, json: { code: 0, data: { children: [{ block_id: 'created_ordered_1' }] } } };
-        }
-
-        expect(body).toEqual({
-          index: 2,
-          children: [
-            {
-              block_type: 2,
-              text: { elements: [{ text_run: { content: 'tail' } }] },
-            },
-          ],
-        });
-        rootCreateCount += 1;
-        return { status: 200, json: { code: 0, data: { children: [{ block_id: 'created_top_1' }] } } };
-      }
-      if (url.includes('/blocks/doc_token_456/children/batch_delete') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('/drive/v1/files/temp_doc_token?type=docx') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('/drive/v1/files/temp_file_token?type=file') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('permissions') && url.includes('transfer_owner')) {
-        return { status: 200, json: { code: 0, msg: 'success' } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# test-note\nUpdated content.',
-    });
-
-    expect(result.docToken).toBe('doc_token_456');
-    expect(rootCreateCount).toBe(2);
-    expect(obsidianMock.requestUrl.mock.calls.some((call) => call[0].url.includes('transfer_owner'))).toBe(false);
-  });
-
-  it('should transfer ownership only for newly created documents', async () => {
-    settings.userId = 'ou-user-123';
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Test Note\nSome text.',
-    });
-
-    expect(result.docToken).toBe('doc_token_456');
-    expect(obsidianMock.requestUrl.mock.calls.some((call) => call[0].url.includes('transfer_owner'))).toBe(true);
-  });
-
-  it('should keep old document content when smart update insertion schema is rejected', async () => {
-    settings.uploadHistory = [{
-      title: 'test-note',
-      url: 'https://feishu.cn/docx/doc_token_456',
-      docToken: 'doc_token_456',
-      sourcePath: 'notes/test-note.md',
-      uploadTime: '2026-06-19T00:00:00Z',
-    }];
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('blocks?page_size')) {
-        if (url.includes('/documents/doc_token_456/blocks?page_size')) {
-          return {
-            status: 200,
-            json: {
-              code: 0,
-              data: {
-                items: [
-                  { block_id: 'doc_token_456', parent_id: '', block_type: 1, children: ['block_child_1'] },
-                  { block_id: 'block_child_1', parent_id: 'doc_token_456', block_type: 2, text: {} },
-                ],
-              },
-            },
-          };
-        }
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'temp_doc_token', parent_id: '', block_type: 1, children: ['temp_paragraph_1'] },
-                { block_id: 'temp_paragraph_1', parent_id: 'temp_doc_token', block_type: 2, text: { elements: [{ text_run: { content: 'Updated content.' } }] } },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks') && options.method === 'POST') {
-        return { status: 200, json: { code: 0, data: { ticket: 'ticket_temp_update' } } };
-      }
-      if (url.includes('/import_tasks/ticket_temp_update')) {
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'temp_doc_token', url: 'https://feishu.cn/docx/temp_doc_token' } } } };
-      }
-      if (url.includes('/documents/doc_token_456/blocks/doc_token_456/children?document_revision_id=-1')) {
-        return {
-          status: 400,
-          json: { code: 1770006, msg: 'schema mismatch' },
-          text: '',
-        };
-      }
-      if (url.includes('/drive/v1/files/temp_doc_token?type=docx') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('/drive/v1/files/temp_file_token?type=file') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('/blocks/doc_token_456/children/batch_delete')) {
-        throw new Error('old blocks must not be deleted before new blocks are inserted successfully');
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const warnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {});
-    await expect(syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# test-note\nUpdated content.',
-    })).rejects.toThrow('schema mismatch');
-
-    const calls = obsidianMock.requestUrl.mock.calls.map((call) => call[0]);
-    expect(calls.some((request) => request.url.includes('/children?document_revision_id=-1'))).toBe(true);
-    expect(calls.some((request) => request.url.includes('children/batch_delete'))).toBe(false);
-    expect(settings.uploadHistory[0].docToken).toBe('doc_token_456');
-    expect(warnSpy).toHaveBeenCalledWith(
-      '[飞书同步] 智能覆盖写入失败，旧文档内容已保留:',
-      expect.any(Error)
-    );
-    warnSpy.mockRestore();
-  });
-
-  it('should not create a duplicate document when Feishu rejects nested children with invalid parameter 9499', async () => {
-    settings.uploadHistory = [{
-      title: 'test-note',
-      url: 'https://feishu.cn/docx/doc_token_456',
-      docToken: 'doc_token_456',
-      sourcePath: 'notes/test-note.md',
-      uploadTime: '2026-06-19T00:00:00Z',
-    }];
-
-    let markdownUploadCount = 0;
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('/documents/doc_token_456/blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'doc_token_456', parent_id: '', block_type: 1, children: ['old_child_1'] },
-                { block_id: 'old_child_1', parent_id: 'doc_token_456', block_type: 2, text: {} },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('/documents/temp_doc_token/blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'temp_doc_token', parent_id: '', block_type: 1, children: ['temp_paragraph_1'] },
-                {
-                  block_id: 'temp_paragraph_1',
-                  parent_id: 'temp_doc_token',
-                  block_type: 2,
-                  text: { elements: [{ text_run: { content: 'Updated content.' } }] },
-                },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('files/upload_all')) {
-        markdownUploadCount += 1;
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks') && options.method === 'POST') {
-        return { status: 200, json: { code: 0, data: { ticket: 'ticket_temp_update' } } };
-      }
-      if (url.includes('/import_tasks/ticket_temp_update')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              result: {
-                job_status: 0,
-                token: 'temp_doc_token',
-                url: 'https://feishu.cn/docx/temp_doc_token',
-              },
-            },
-          },
-        };
-      }
-      if (url.includes('/documents/doc_token_456/blocks/doc_token_456/children?document_revision_id=-1')) {
-        return {
-          status: 400,
-          json: {
-            code: 9499,
-            msg: 'Invalid parameter type in json: children. Invalid parameter value: {"block_type":12,"bullet":{"elements":[{"text_run":{"content":"清洗数据 (Python)"}}]}}. Please check and modify accordingly.',
-          },
-          text: '',
-        };
-      }
-      if (url.includes('/drive/v1/files/temp_doc_token?type=docx') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('/drive/v1/files/temp_file_token?type=file') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('/blocks/doc_token_456/children/batch_delete')) {
-        throw new Error('old blocks must not be deleted when new block insertion is rejected');
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const warnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {});
-    await expect(syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# test-note\nUpdated content.',
-    })).rejects.toThrow('Invalid parameter type in json: children');
-
-    const calls = obsidianMock.requestUrl.mock.calls.map((call) => call[0]);
-    expect(markdownUploadCount).toBe(1);
-    expect(calls.some((request) => request.url.includes('children/batch_delete'))).toBe(false);
-    expect(calls.some((request) => request.url.includes('/drive/v1/files/temp_doc_token?type=docx') && request.method === 'DELETE')).toBe(true);
-    expect(settings.uploadHistory[0].docToken).toBe('doc_token_456');
-    expect(warnSpy).toHaveBeenCalledWith(
-      '[飞书同步] 智能覆盖写入失败，旧文档内容已保留:',
-      expect.any(Error)
-    );
-    expect(warnSpy).not.toHaveBeenCalledWith(
-      '[飞书同步] 智能覆盖更新失败，降级为新建文档:',
-      expect.any(Error)
-    );
-    warnSpy.mockRestore();
-  });
-
-  it('should relink a stale history token from folder and continue smart update', async () => {
-    settings.uploadHistory = [{
-      title: 'test-note',
-      url: 'https://feishu.cn/docx/stale_doc_token',
-      docToken: 'stale_doc_token',
-      sourcePath: 'notes/test-note.md',
-      uploadTime: '2026-06-19T00:00:00Z',
-    }];
-
-    const warnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {});
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files?folder_token')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              files: [
-                { type: 'docx', name: 'test-note', token: 'relinked_doc_token' },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('/documents/stale_doc_token/blocks?page_size')) {
-        return {
-          status: 400,
-          json: { code: 1770003, msg: 'resource deleted' },
-          text: '',
-        };
-      }
-      if (url.includes('/documents/relinked_doc_token/blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'relinked_doc_token', parent_id: '', block_type: 1, children: ['relinked_child_1'] },
-                { block_id: 'relinked_child_1', parent_id: 'relinked_doc_token', block_type: 2, text: { elements: [{ text_run: { content: 'old content' } }] } },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('/blocks/relinked_doc_token/children/batch_delete') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks') && options.method === 'POST') {
-        return { status: 200, json: { code: 0, data: { ticket: 'ticket_temp_update' } } };
-      }
-      if (url.includes('/import_tasks/ticket_temp_update')) {
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'temp_doc_token', url: 'https://feishu.cn/docx/temp_doc_token' } } } };
-      }
-      if (url.includes('/documents/temp_doc_token/blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'temp_doc_token', parent_id: '', block_type: 1, children: ['temp_paragraph_1'] },
-                { block_id: 'temp_paragraph_1', parent_id: 'temp_doc_token', block_type: 2, text: { elements: [{ text_run: { content: 'Updated content.' } }] } },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('/documents/relinked_doc_token/blocks/relinked_doc_token/children?document_revision_id=-1')) {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('/drive/v1/files/temp_doc_token?type=docx') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('/drive/v1/files/temp_file_token?type=file') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('permissions') && url.includes('transfer_owner')) {
-        return { status: 200, json: { code: 0, msg: 'success' } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# test-note\nUpdated content.',
-    });
-
-    expect(result.docToken).toBe('relinked_doc_token');
-    expect(result.url).toBe('https://open.feishu.cn/docx/relinked_doc_token');
-    expect(settings.uploadHistory[0]).toMatchObject({
-      docToken: 'relinked_doc_token',
-      sourcePath: 'notes/test-note.md',
-    });
-    expect(warnSpy).toHaveBeenCalledWith('[飞书同步] 检测到历史飞书 token 已失效，已清理本地关联记录');
-    warnSpy.mockRestore();
-  });
-
-  it('should keep the imported document when image block scanning fails after import', async () => {
-    const localFile = {
-      path: 'notes/attachments/local.png',
-      name: 'local.png',
-      extension: 'png',
-      bytes: makePngBytes(640, 320),
-    };
-    app.metadataCache.getFirstLinkpathDest.mockImplementation((linkpath) => (
-      linkpath === 'attachments/local.png' ? localFile : null
-    ));
-    app.vault.getAbstractFileByPath = vi.fn((filePath) => (
-      filePath === 'notes/attachments/local.png' ? localFile : null
-    ));
-    app.vault.getResourcePath = vi.fn(() => 'app://local/local.png');
-    app.vault.readBinary.mockImplementation(async (file) => file.bytes);
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { status: 200, json: { code: 0, data: { ticket: 'ticket_123' } } };
-        }
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'doc_token_456', url: 'https://feishu.cn/docx/doc_token_456' } } } };
-      }
-      if (url.includes('blocks?page_size')) {
-        return {
-          status: 404,
-          json: { code: 404, msg: 'document block not found' },
-          text: '',
-        };
-      }
-      if (url.includes('permissions') && url.includes('transfer_owner')) {
-        return { status: 200, json: { code: 0, msg: 'success' } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const warnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {});
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Test Note\n![Local](attachments/local.png)',
-    });
-
-    expect(result.docToken).toBe('doc_token_456');
-    expect(result.url).toBe('https://feishu.cn/docx/doc_token_456');
-    expect(settings.uploadHistory.length).toBe(1);
-    expect(warnSpy).toHaveBeenCalledWith(
-      '[飞书同步] 图片后处理跳过，文档正文已导入:',
-      expect.any(Error)
-    );
-    expect(result.imageSummary.failed).toBe(1);
-    warnSpy.mockRestore();
-  });
-
-  it('should upload prepared local image assets and replace Feishu image blocks', async () => {
-    const localFile = {
-      path: 'notes/attachments/local.png',
-      name: 'local.png',
-      extension: 'png',
-      bytes: makePngBytes(100, 50),
-    };
-    app.metadataCache.getFirstLinkpathDest.mockImplementation((linkpath) => (
-      linkpath === 'attachments/local.png' ? localFile : null
-    ));
-    app.vault.getAbstractFileByPath = vi.fn((filePath) => (
-      filePath === 'notes/attachments/local.png' ? localFile : null
-    ));
-    app.vault.getResourcePath = vi.fn(() => 'app://local/local.png');
-    app.vault.readBinary.mockImplementation(async (file) => file.bytes);
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { status: 200, json: { code: 0, data: { ticket: 'ticket_123' } } };
-        }
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'doc_token_456', url: 'https://feishu.cn/docx/doc_token_456' } } } };
-      }
-      if (url.includes('blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [{ block_id: 'image_block_1', parent_id: 'doc_token_456', block_type: 27 }],
-            },
-          },
-        };
-      }
-      if (url.includes('medias/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'image_token_1' } } };
-      }
-      if (url.includes('/blocks/image_block_1')) {
-        return { status: 200, json: { code: 0, data: {} } };
-      }
-      if (url.includes('permissions') && url.includes('transfer_owner')) {
-        return { status: 200, json: { code: 0, msg: 'success' } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Test Note\n![Local](attachments/local.png)',
-    });
-
-    expect(result.docToken).toBe('doc_token_456');
-    expect(result.imageSummary).toMatchObject({
-      uploaded: 1,
-      skipped: 0,
-      failed: 0,
-    });
-    const calls = obsidianMock.requestUrl.mock.calls.map((call) => call[0]);
-    const markdownUpload = calls.find((request) => request.url.includes('files/upload_all'));
-    const markdownBody = new TextDecoder().decode(new Uint8Array(markdownUpload.body));
-    expect(markdownBody).toContain('![Local](https://obsidian-wechat-converter.invalid/feishu-local-image/image-1.png)');
-    expect(calls.some((request) => request.url.includes('medias/upload_all'))).toBe(true);
-    expect(calls.some((request) => request.url.includes('/blocks/image_block_1') && request.method === 'PATCH')).toBe(true);
-    const imagePatch = calls.find((request) => request.url.includes('/blocks/image_block_1') && request.method === 'PATCH');
-    expect(JSON.parse(imagePatch.body).replace_image).toEqual({
-      token: 'image_token_1',
-      width: 100,
-      height: 50,
-      align: 2,
-    });
-    const imageUpload = calls.find((request) => request.url.includes('medias/upload_all'));
-    const imageUploadBody = new TextDecoder().decode(new Uint8Array(imageUpload.body));
-    expect(imageUploadBody).toContain('Content-Type: image/png');
-  });
-
-  it('should leave remote images to Feishu import without re-uploading them', async () => {
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { status: 200, json: { code: 0, data: { ticket: 'ticket_123' } } };
-        }
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'doc_token_456', url: 'https://feishu.cn/docx/doc_token_456' } } } };
-      }
-      if (url.includes('blocks?page_size')) {
-        throw new Error('remote images should not trigger image block replacement');
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Test Note\n![Remote](https://cdn.example.com/a.png)',
-    });
-
-    expect(result.imageSummary).toEqual({
-      uploaded: 0,
-      skipped: 0,
-      failed: 0,
-      details: [],
-    });
-    const calls = obsidianMock.requestUrl.mock.calls.map((call) => call[0]);
-    expect(calls.some((request) => request.url.includes('medias/upload_all'))).toBe(false);
-  });
-
-  it('should re-upload remote images when smart updating an existing document', async () => {
-    settings.uploadHistory = [{
-      title: 'test-note',
-      url: 'https://feishu.cn/docx/doc_token_456',
-      docToken: 'doc_token_456',
-      sourcePath: 'notes/test-note.md',
-      uploadTime: '2026-06-19T00:00:00Z',
-    }];
-
-    const originalFetch = globalThis.fetch;
-    const fetchSpy = vi.fn();
-    globalThis.fetch = fetchSpy;
-    let didInsertNewBlocks = false;
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('/documents/doc_token_456/blocks?page_size')) {
-        if (didInsertNewBlocks) {
-          return {
-            status: 200,
-            json: {
-              code: 0,
-              data: {
-                items: [
-                  { block_id: 'doc_token_456', parent_id: '', block_type: 1, children: ['created_heading_1', 'created_remote_image_1'] },
-                  { block_id: 'created_heading_1', parent_id: 'doc_token_456', block_type: 3, heading1: { elements: [{ text_run: { content: 'Updated title' } }] } },
-                  { block_id: 'created_remote_image_1', parent_id: 'doc_token_456', block_type: 27, image: { token: '', width: 1303, height: 409 } },
-                ],
-              },
-            },
-          };
-        }
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'doc_token_456', parent_id: '', block_type: 1, children: ['old_child_1'] },
-                { block_id: 'old_child_1', parent_id: 'doc_token_456', block_type: 2, text: { elements: [{ text_run: { content: 'old' } }] } },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('/documents/temp_doc_token/blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'temp_doc_token', parent_id: '', block_type: 1, children: ['temp_heading', 'temp_image'] },
-                { block_id: 'temp_heading', parent_id: 'temp_doc_token', block_type: 3, heading1: { elements: [{ text_run: { content: 'Updated title' } }] } },
-                { block_id: 'temp_image', parent_id: 'temp_doc_token', block_type: 27, image: { token: 'remote_file_token', width: 1303, height: 409, scale: 1 } },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks') && options.method === 'POST') {
-        return { status: 200, json: { code: 0, data: { ticket: 'ticket_temp_update' } } };
-      }
-      if (url.includes('/import_tasks/ticket_temp_update')) {
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'temp_doc_token', url: 'https://feishu.cn/docx/temp_doc_token' } } } };
-      }
-      if (url.includes('/documents/doc_token_456/blocks/doc_token_456/children?document_revision_id=-1') && options.method === 'POST') {
-        const body = JSON.parse(options.body);
-        expect(body).toEqual({
-          index: 1,
-          children: [
-            { block_type: 3, heading1: { elements: [{ text_run: { content: 'Updated title' } }] } },
-            { block_type: 27, image: { file_token: 'remote_file_token', width: 1303, height: 409 } },
-          ],
-        });
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              children: [
-                { block_id: 'created_heading_1', block_type: 3, parent_id: 'doc_token_456' },
-                { block_id: 'created_remote_image_1', block_type: 27, parent_id: 'doc_token_456', image: { token: '', width: 1303, height: 409 } },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('/blocks/doc_token_456/children/batch_delete') && options.method === 'DELETE') {
-        didInsertNewBlocks = true;
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('/drive/v1/files/temp_doc_token?type=docx') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url.includes('/drive/v1/files/temp_file_token?type=file') && options.method === 'DELETE') {
-        return { status: 200, json: { code: 0 } };
-      }
-      if (url === 'https://cdn.example.com/remote.png') {
-        return {
-          status: 200,
-          headers: { 'content-type': 'image/png' },
-          arrayBuffer: makePngBytes(1303, 409),
-        };
-      }
-      if (url.includes('medias/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'reuploaded_remote_image_token' } } };
-      }
-      if (url.includes('/documents/doc_token_456/blocks/created_remote_image_1?document_revision_id=-1') && options.method === 'PATCH') {
-        return { status: 200, json: { code: 0, data: {} } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    try {
-      const result = await syncNoteToFeishu({
-        app,
-        settings,
-        activeFile,
-        markdown: '# test-note\n![bob](https://cdn.example.com/remote.png)',
-      });
-
-      expect(result.docToken).toBe('doc_token_456');
-      expect(result.imageSummary).toMatchObject({
-        uploaded: 1,
-        skipped: 0,
-        failed: 0,
-      });
-      expect(fetchSpy).not.toHaveBeenCalled();
-      const calls = obsidianMock.requestUrl.mock.calls.map((call) => call[0]);
-      expect(calls.some((request) => request.url === 'https://cdn.example.com/remote.png')).toBe(true);
-      expect(calls.some((request) => request.url.includes('medias/upload_all'))).toBe(true);
-      const imagePatch = calls.find((request) => request.url.includes('/documents/doc_token_456/blocks/created_remote_image_1?document_revision_id=-1') && request.method === 'PATCH');
-      expect(JSON.parse(imagePatch.body).replace_image).toEqual({
-        token: 'reuploaded_remote_image_token',
-        width: 1303,
-        height: 409,
-        align: 2,
-      });
-    } finally {
-      globalThis.fetch = originalFetch;
-    }
-  });
-
-  it('should replace local images at their original markdown image block positions', async () => {
-    const localFile = {
-      path: 'notes/attachments/local.png',
-      name: 'local.png',
-      extension: 'png',
-      bytes: makePngBytes(640, 320),
-    };
-    app.metadataCache.getFirstLinkpathDest.mockImplementation((linkpath) => (
-      linkpath === 'attachments/local.png' ? localFile : null
-    ));
-    app.vault.getAbstractFileByPath = vi.fn((filePath) => (
-      filePath === 'notes/attachments/local.png' ? localFile : null
-    ));
-    app.vault.getResourcePath = vi.fn(() => 'app://local/local.png');
-    app.vault.readBinary.mockImplementation(async (file) => file.bytes);
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { status: 200, json: { code: 0, data: { ticket: 'ticket_123' } } };
-        }
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'doc_token_456', url: 'https://feishu.cn/docx/doc_token_456' } } } };
-      }
-      if (url.includes('blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'remote_image_block', parent_id: 'doc_token_456', block_type: 27 },
-                { block_id: 'local_image_block', parent_id: 'doc_token_456', block_type: 27 },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('medias/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'image_token_1' } } };
-      }
-      if (url.includes('/blocks/local_image_block')) {
-        return { status: 200, json: { code: 0, data: {} } };
-      }
-      if (url.includes('/blocks/remote_image_block')) {
-        throw new Error('remote image block must not be replaced by local image upload');
-      }
-      if (url.includes('permissions') && url.includes('transfer_owner')) {
-        return { status: 200, json: { code: 0, msg: 'success' } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: [
-        '# Test Note',
-        '![Remote](https://cdn.example.com/a.png)',
-        '![Local](attachments/local.png)',
-      ].join('\n'),
-    });
-
-    expect(result.imageSummary).toMatchObject({
-      uploaded: 1,
-      skipped: 0,
-      failed: 0,
-    });
-    const calls = obsidianMock.requestUrl.mock.calls.map((call) => call[0]);
-    expect(calls.some((request) => request.url.includes('/blocks/remote_image_block'))).toBe(false);
-    expect(calls.some((request) => request.url.includes('/blocks/local_image_block') && request.method === 'PATCH')).toBe(true);
-    const imagePatch = calls.find((request) => request.url.includes('/blocks/local_image_block') && request.method === 'PATCH');
-    expect(JSON.parse(imagePatch.body).replace_image).toEqual({
-      token: 'image_token_1',
-      width: 640,
-      height: 320,
-      align: 2,
-    });
-  });
-
-  it('should keep wiki image and relative image replacements aligned after a remote image', async () => {
-    const wikiFile = {
-      path: 'notes/attachments/音乐卡点调整.png',
-      name: '音乐卡点调整.png',
-      extension: 'png',
-      bytes: makePngBytes(1000, 500),
-    };
-    const relativeFile = {
-      path: 'notes/attachments/打工.png',
-      name: '打工.png',
-      extension: 'png',
-      bytes: makePngBytes(800, 400),
-    };
-    const gifFile = {
-      path: 'notes/测试.gif',
-      name: '测试.gif',
-      extension: 'gif',
-      bytes: makeGifBytes(480, 270),
-    };
-    app.metadataCache.getFirstLinkpathDest.mockImplementation((linkpath) => {
-      if (linkpath === 'attachments/音乐卡点调整.png') return wikiFile;
-      if (linkpath === 'attachments/打工.png') return relativeFile;
-      if (linkpath === '测试.gif') return gifFile;
-      return null;
-    });
-    app.vault.getAbstractFileByPath = vi.fn((filePath) => {
-      if (filePath === 'notes/attachments/音乐卡点调整.png') return wikiFile;
-      if (filePath === 'notes/attachments/打工.png') return relativeFile;
-      if (filePath === 'notes/测试.gif') return gifFile;
-      return null;
-    });
-    app.vault.getResourcePath = vi.fn((file) => `app://local/${encodeURIComponent(file.path)}`);
-    app.vault.readBinary.mockImplementation(async (file) => file.bytes);
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { status: 200, json: { code: 0, data: { ticket: 'ticket_123' } } };
-        }
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'doc_token_456', url: 'https://feishu.cn/docx/doc_token_456' } } } };
-      }
-      if (url.includes('blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'remote_image_block', parent_id: 'doc_token_456', block_type: 27 },
-                { block_id: 'wiki_image_block', parent_id: 'doc_token_456', block_type: 27 },
-                { block_id: 'relative_image_block', parent_id: 'doc_token_456', block_type: 27 },
-                { block_id: 'gif_image_block', parent_id: 'doc_token_456', block_type: 27 },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('medias/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: url.includes('token-two') ? 'image_token_2' : 'image_token_1' } } };
-      }
-      if (url.includes('/blocks/wiki_image_block') || url.includes('/blocks/relative_image_block') || url.includes('/blocks/gif_image_block')) {
-        return { status: 200, json: { code: 0, data: {} } };
-      }
-      if (url.includes('/blocks/remote_image_block')) {
-        throw new Error('non-local prepared image block must not be replaced');
-      }
-      if (url.includes('permissions') && url.includes('transfer_owner')) {
-        return { status: 200, json: { code: 0, msg: 'success' } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: [
-        '# Test Note',
-        '![Remote](https://cdn.example.com/a.png)',
-        '![[attachments/音乐卡点调整.png|音乐|510]]',
-        '![测试](attachments/打工.png)',
-        '![不对](测试.gif)',
-      ].join('\n'),
-    });
-
-    expect(result.imageSummary).toMatchObject({
-      uploaded: 3,
-      skipped: 0,
-      failed: 0,
-    });
-    const calls = obsidianMock.requestUrl.mock.calls.map((call) => call[0]);
-    const markdownUpload = calls.find((request) => request.url.includes('files/upload_all'));
-    const markdownBody = new TextDecoder().decode(new Uint8Array(markdownUpload.body));
-    expect(markdownBody).toContain('![音乐](https://obsidian-wechat-converter.invalid/feishu-local-image/image-1.png)');
-    expect(markdownBody).toContain('![测试](https://obsidian-wechat-converter.invalid/feishu-local-image/image-2.png)');
-    expect(markdownBody).toContain('![不对](https://obsidian-wechat-converter.invalid/feishu-local-image/image-3.gif)');
-    expect(calls.some((request) => request.url.includes('/blocks/remote_image_block'))).toBe(false);
-    expect(calls.some((request) => request.url.includes('/blocks/wiki_image_block') && request.method === 'PATCH')).toBe(true);
-    expect(calls.some((request) => request.url.includes('/blocks/relative_image_block') && request.method === 'PATCH')).toBe(true);
-    expect(calls.some((request) => request.url.includes('/blocks/gif_image_block') && request.method === 'PATCH')).toBe(true);
-    const wikiPatch = calls.find((request) => request.url.includes('/blocks/wiki_image_block') && request.method === 'PATCH');
-    expect(JSON.parse(wikiPatch.body).replace_image).toEqual({
-      token: 'image_token_1',
-      width: 1000,
-      height: 500,
-      align: 2,
-    });
-    const relativePatch = calls.find((request) => request.url.includes('/blocks/relative_image_block') && request.method === 'PATCH');
-    expect(JSON.parse(relativePatch.body).replace_image).toEqual({
-      token: 'image_token_1',
-      width: 800,
-      height: 400,
-      align: 2,
-    });
-    const gifPatch = calls.find((request) => request.url.includes('/blocks/gif_image_block') && request.method === 'PATCH');
-    expect(JSON.parse(gifPatch.body).replace_image).toEqual({
-      token: 'image_token_1',
-      width: 480,
-      height: 270,
-      align: 2,
-    });
-  });
-
-  it('should report Feishu image progress without exposing filenames', async () => {
-    const localFile = {
-      path: 'notes/attachments/local.png',
-      name: 'local.png',
-      extension: 'png',
-      bytes: makePngBytes(640, 320),
-    };
-    app.metadataCache.getFirstLinkpathDest.mockImplementation((linkpath) => (
-      linkpath === 'attachments/local.png' ? localFile : null
-    ));
-    app.vault.getAbstractFileByPath = vi.fn((filePath) => (
-      filePath === 'notes/attachments/local.png' ? localFile : null
-    ));
-    app.vault.getResourcePath = vi.fn(() => 'app://local/local.png');
-    app.vault.readBinary.mockImplementation(async (file) => file.bytes);
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { status: 200, json: { code: 0, data: { ticket: 'ticket_123' } } };
-        }
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'doc_token_456', url: 'https://feishu.cn/docx/doc_token_456' } } } };
-      }
-      if (url.includes('blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [{ block_id: 'image_block_1', parent_id: 'doc_token_456', block_type: 27 }],
-            },
-          },
-        };
-      }
-      if (url.includes('medias/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'image_token_1' } } };
-      }
-      if (url.includes('/blocks/image_block_1')) {
-        return { status: 200, json: { code: 0, data: {} } };
-      }
-      if (url.includes('permissions') && url.includes('transfer_owner')) {
-        return { status: 200, json: { code: 0, msg: 'success' } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const progressMessages = [];
-    await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Test Note\n![Local|320](attachments/local.png)',
-      onProgress: (_stage, message) => {
-        progressMessages.push(message);
-      },
-    });
-
-    expect(progressMessages).toContain('正在同步正文图片 (1/1)...');
-    expect(progressMessages.some((message) => String(message).includes('local.png'))).toBe(false);
-  });
-
-  it('should upload local GIF placeholders during Feishu image post-processing', async () => {
-    const gifFile = {
-      path: 'notes/attachments/demo.gif',
-      name: 'demo.gif',
-      extension: 'gif',
-      bytes: makeGifBytes(300, 200),
-    };
-    app.metadataCache.getFirstLinkpathDest.mockReturnValue(gifFile);
-    app.vault.getAbstractFileByPath = vi.fn(() => gifFile);
-    app.vault.getResourcePath = vi.fn(() => 'app://local/demo.gif');
-    app.vault.readBinary.mockImplementation(async (file) => file.bytes);
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { status: 200, json: { code: 0, data: { ticket: 'ticket_123' } } };
-        }
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'doc_token_456', url: 'https://feishu.cn/docx/doc_token_456' } } } };
-      }
-      if (url.includes('blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [{ block_id: 'image_block_1', parent_id: 'doc_token_456', block_type: 27 }],
-            },
-          },
-        };
-      }
-      if (url.includes('medias/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'image_token_1' } } };
-      }
-      if (url.includes('/blocks/image_block_1')) {
-        return { status: 200, json: { code: 0, data: {} } };
-      }
-      if (url.includes('permissions') && url.includes('transfer_owner')) {
-        return { status: 200, json: { code: 0, msg: 'success' } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const errorSpy = vi.spyOn(console, 'error').mockImplementation(() => {});
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Test Note\n![Gif](attachments/demo.gif)',
-    });
-
-    expect(result.docToken).toBe('doc_token_456');
-    expect(result.imageSummary.uploaded).toBe(1);
-    expect(result.imageSummary.skipped).toBe(0);
-    expect(result.imageSummary.failed).toBe(0);
-    expect(app.vault.readBinary).toHaveBeenCalled();
-    const calls = obsidianMock.requestUrl.mock.calls.map((call) => call[0]);
-    expect(calls.some((request) => request.url.includes('medias/upload_all'))).toBe(true);
-    expect(errorSpy).not.toHaveBeenCalled();
-    errorSpy.mockRestore();
-  });
-
-  it('should prefer vault bytes over asset base64 when uploading local GIFs', async () => {
-    const gifFile = {
-      path: 'notes/attachments/demo.gif',
-      name: 'demo.gif',
-      extension: 'gif',
-      bytes: makeGifBytes(320, 180),
-    };
-    app.metadataCache.getFirstLinkpathDest.mockReturnValue(gifFile);
-    app.vault.getAbstractFileByPath = vi.fn(() => gifFile);
-    app.vault.getResourcePath = vi.fn(() => 'app://local/demo.gif');
-    app.vault.readBinary.mockImplementation(async (file) => file.bytes);
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { status: 200, json: { code: 0, data: { ticket: 'ticket_123' } } };
-        }
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'doc_token_456', url: 'https://feishu.cn/docx/doc_token_456' } } } };
-      }
-      if (url.includes('blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [{ block_id: 'image_block_1', parent_id: 'doc_token_456', block_type: 27 }],
-            },
-          },
-        };
-      }
-      if (url.includes('medias/upload_all')) {
-        const bodyText = new TextDecoder().decode(new Uint8Array(options.body));
-        expect(bodyText).toContain('filename="demo.gif"');
-        expect(bodyText).not.toContain('not-a-real-gif');
-        return { status: 200, json: { code: 0, data: { file_token: 'image_token_1' } } };
-      }
-      if (url.includes('/blocks/image_block_1')) {
-        return { status: 200, json: { code: 0, data: {} } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# Test Note\n![Gif](attachments/demo.gif)',
-      renderMermaidFenceToDataUrl: async () => 'data:image/png;base64,bm90LWEtcmVhbC1naWY=',
-    });
-
-    expect(result.imageSummary).toMatchObject({
-      uploaded: 1,
-      failed: 0,
-    });
-    expect(app.vault.readBinary).toHaveBeenCalledTimes(2);
-  });
-
-  it('should fall back to creating a new document when smart update deletion fails', async () => {
-    settings.uploadHistory = [{
-      title: 'test-note',
-      url: 'https://feishu.cn/docx/old_doc_token',
-      docToken: 'old_doc_token',
-      sourcePath: 'notes/test-note.md',
-      uploadTime: '2026-06-19T00:00:00Z',
-    }];
-
-    obsidianMock.requestUrl.mockImplementation(async (options) => {
-      const url = options.url || '';
-      if (url.includes('tenant_access_token')) {
-        return { status: 200, json: { code: 0, tenant_access_token: 't-123', expire: 7200 } };
-      }
-      if (url.includes('blocks?page_size')) {
-        return {
-          status: 200,
-          json: {
-            code: 0,
-            data: {
-              items: [
-                { block_id: 'old_child_1', parent_id: 'old_doc_token', block_type: 1 },
-              ],
-            },
-          },
-        };
-      }
-      if (url.includes('/blocks/old_doc_token/children/batch_delete') && options.method === 'DELETE') {
-        return {
-          status: 404,
-          text: '404 page not found',
-        };
-      }
-      if (url.includes('files/upload_all')) {
-        return { status: 200, json: { code: 0, data: { file_token: 'temp_file_token' } } };
-      }
-      if (url.includes('import_tasks')) {
-        if (options.method === 'POST') {
-          return { status: 200, json: { code: 0, data: { ticket: 'ticket_123' } } };
-        }
-        return { status: 200, json: { code: 0, data: { result: { job_status: 0, token: 'new_doc_token', url: 'https://feishu.cn/docx/new_doc_token' } } } };
-      }
-      if (url.includes('permissions') && url.includes('transfer_owner')) {
-        return { status: 200, json: { code: 0, msg: 'success' } };
-      }
-      return { status: 200, json: { code: 0 } };
-    });
-
-    const warnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {});
-    const result = await syncNoteToFeishu({
-      app,
-      settings,
-      activeFile,
-      markdown: '# test-note\nUpdated content.',
-    });
-
-    expect(result.docToken).toBe('new_doc_token');
-    expect(result.url).toBe('https://feishu.cn/docx/new_doc_token');
-    expect(settings.uploadHistory[0].docToken).toBe('new_doc_token');
-    expect(warnSpy).toHaveBeenCalledWith(
-      '[飞书同步] 智能覆盖更新失败，降级为新建文档:',
-      expect.any(Error)
-    );
-    warnSpy.mockRestore();
   });
 });
